Place profile menu above toasts
