+++ play/src/front/Components/Util/FloatingUiPopupList.svelte
@@ -13,7 +13,7 @@
     {@const Component = value.componentType}
     {@const action = value.action}
     {@const arrowAction = value.arrowAction}
-    <div class="absolute z-[3000]" use:action>
+    <div class="absolute z-[3000]" style:z-index={value.zIndex} use:action>
         {#if arrowAction}
             <div use:arrowAction></div>
         {/if}
